Add files via upload
For this class project I decided I wanted to create a database for an e-commerce shop. The shop would be a simple product reseller, but I wanted to add some functionality to it that I had not seen in the coursework. I determined that the most useful addition would be an inventory system that tracks the product and manufacturer. This data would be carried through an orderline which allows the database to track what product, from which manufacturer is being sold. That way the database would know how much of a product is sold and remains in inventory, which would enable the shop to restock inventory and inform product purchase decisions.